feat(eslint-plugin): add stable-query-client rule (#6077)
* feat: add stable-query-client rule

This commit introduces a new rule in the eslint-plugin-query package. The rule, named 'stable-query-client', ensures that the QueryClient is
stable. It checks if the QueryClient is either extracted from the component or wrapped in React.useState. If not, it suggests a fix to wrap
the QueryClient instantiation in a React.useState call. This rule is recommended to be set as an error.

* update recommended snapshot

* fix eslint

* check only immediate instantiation

* fix:  allow QueryClient invocation in react server components

Updated the stable-query-client rule to allow the invocation of QueryClient inside react server components.

* docs: stable-query-client

---------

Co-authored-by: Dominik Dorfmeister <office@dorfmeister.cc>